Set default mode for callgraph to 'true' in the plugin.xml.
Have cancelJob and isJobCancelled set/check a flag jobCancelled in SystemTapParser.

+++ systemtap/org.eclipse.linuxtools.callgraph.launch/src/org/eclipse/linuxtools/callgraph/launch/SystemTapLaunchConfigurationDelegate.java
@@ -338,9 +338,11 @@ public class SystemTapLaunchConfigurationDelegate extends
 			while (!process.isTerminated()) {
 				Thread.sleep(100);
 				if ((monitor != null && monitor.isCanceled()) || parser.isJobCancelled()) {
-					parser.cancelJob();
-					Runtime run = Runtime.getRuntime();
-					run.exec("kill stap"); //$NON-NLS-1$
+					if (!parser.isJobCancelled()) {
+						parser.cancelJob();
+						Runtime run = Runtime.getRuntime();
+						run.exec("kill stap"); //$NON-NLS-1$
+					}
 					process.terminate();
 					return;
 				}
